+++ covalent_dispatcher/_core/dispatcher.py
@@ -31,7 +31,7 @@ from covalent._shared_files.defaults import WAIT_EDGE_NAME, parameter_prefix
 from covalent._shared_files.util_classes import RESULT_STATUS
 
 from . import data_manager as datasvc
-from . import runner
+from . import runner_ng
 from .data_modules import graph as tg_utils
 from .data_modules import job_manager as jbmgr
 from .dispatcher_modules.caches import _pending_parents, _sorted_task_groups, _unresolved_tasks
@@ -224,19 +224,11 @@ async def _submit_task_group(dispatch_id: str, sorted_nodes: List[int], task_gro
             app_log.debug(f"Using new runner for task group {task_group_id}")
 
             known_nodes = list(set(known_nodes))
-
-            task_spec = task_specs[0]
-            abstract_inputs = {"args": task_spec["args_ids"], "kwargs": task_spec["kwargs_ids"]}
-
-            # Temporarily redirect to in-memory runner (this is incompatible with task packing)
-            if len(task_specs) > 1:
-                raise RuntimeError("Task packing is not supported yet.")
-
-            coro = runner.run_abstract_task(
+            coro = runner_ng.run_abstract_task_group(
                 dispatch_id=dispatch_id,
-                node_id=task_group_id,
-                node_name=node_name,
-                abstract_inputs=abstract_inputs,
+                task_group_id=task_group_id,
+                task_seq=task_specs,
+                known_nodes=known_nodes,
                 selected_executor=[selected_executor, selected_executor_data],
             )